// src/api/generated/model/error.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * mdrive API
 * POSIX-like file system API over S3
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorCode } from './errorCode';

export interface Error {
  /** Machine-readable error code */
  code: ErrorCode;
  /** Human-readable error message */
  message: string;
}